hotema-backend/requirements.txt
Django==5.2.4
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-decouple
psycopg2-binary
